requirements.txt
aiohttp==3.9.0
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.25.0
anyio==4.11.0
attrs==25.4.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
distro==1.9.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
google-auth==2.43.0
google-genai==1.49.0
greenlet==3.0.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.2
idna==3.11
jiter==0.12.0
multidict==6.7.0
openai==2.8.0
packaging==25.0
Pillow==10.1.0
playwright==1.40.0
propcache==0.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
pyee==11.0.1
python-dotenv==1.0.0
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
shellingham==1.5.4
sniffio==1.3.1
tenacity==9.1.2
tokenizers==0.22.1
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
yarl==1.22.0
